whisper-service/requirements.txt
flask==3.0.0
openai-whisper==20250625
